Crate wasmedge_macro

Overview

The wasmedge-macro crate defines a group of procedural macros used by both wasmedge-sdk and wasmedge-sys crates.

Attribute Macros

async_host_function

Declare a native async function that will be used to create an async host function instance.

host_function

Declare a native function that will be used to create a host function instance.